Revert "controller fix pull request port because yes"
This reverts commit b6f2239a864fe1bae8c7a15731bb6c73bf54e1e8.

+++ RSDKv4/Input.hpp
@@ -104,8 +104,8 @@ enum ExtraSDLButtons {
     SDL_CONTROLLER_BUTTON_MAX_EXTRA,
 };
 
-void controllerInit(int controllerID);
-void controllerClose(int controllerID);
+void controllerInit(byte controllerID);
+void controllerClose(byte controllerID);
 #endif
 
 #if RETRO_USING_SDL1
